requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiogram_calendar==0.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
environs==14.1.1
flake8==7.2.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
isort==6.0.1
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
multidict==6.2.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
propcache==0.3.1
psycopg==3.2.6
pycodestyle==2.13.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
tzlocal==5.3.1
yarl==1.18.3
